Moved all redis functionality
